Art assets
Art Assets are crudely in, re-ising and font changes still to be done, and the hints need to be re-adjusted due to the lack of art assets in them now, but beyond that most, if not all art should be stylized with custom art. There is also currently a bug with the background not changing, I will fix this tomorrow or later today.